requirements.txt
python-telegram-bot
google-generativeai
supabase
requests
python-dotenv
schedule
rapidfuzz